+++ src/test/java/eu/senla/LogoutTest.java
@@ -10,31 +10,31 @@ import org.junit.jupiter.api.Test;
 import org.openqa.selenium.By;
 
 public class LogoutTest extends BaseTest {
-
-  @DisplayName("Проверка успешного выхода из учетной записи")
-  @Tag("smoke")
-  @Test
-  public void testLogout() {
-    successfulLogin();
-
-    driver.findElement(By.cssSelector("span.oxd-userdropdown-tab")).click();
-    driver.findElement(By.cssSelector("a[href='/web/index.php/auth/logout']")).click();
-
-    wait.until(d -> driver.findElement(By.xpath("//button[text()=' Login ']")).isDisplayed());
-
-    assertAll(
-        () ->
-            assertTrue(
-                driver.findElement(By.xpath("//button[text()=' Login ']")).isDisplayed(),
-                "Unsuccessful logout"),
-        () ->
-            assertTrue(
-                driver.findElement(By.xpath("//button[text()=' Login ']")).isDisplayed(),
-                "Unsuccessful logout"),
-        () ->
-            assertEquals(
-                "https://opensource-demo.orangehrmlive.com/web/index.php/auth/login",
-                driver.getCurrentUrl(),
-                "URL doesn't match"));
-  }
+//
+//  @DisplayName("Проверка успешного выхода из учетной записи")
+//  @Tag("smoke")
+//  @Test
+//  public void testLogout() {
+//    successfulLogin();
+//
+//    driver.findElement(By.cssSelector("span.oxd-userdropdown-tab")).click();
+//    driver.findElement(By.cssSelector("a[href='/web/index.php/auth/logout']")).click();
+//
+//    wait.until(d -> driver.findElement(By.xpath("//button[text()=' Login ']")).isDisplayed());
+//
+//    assertAll(
+//        () ->
+//            assertTrue(
+//                driver.findElement(By.xpath("//button[text()=' Login ']")).isDisplayed(),
+//                "Unsuccessful logout"),
+//        () ->
+//            assertTrue(
+//                driver.findElement(By.xpath("//button[text()=' Login ']")).isDisplayed(),
+//                "Unsuccessful logout"),
+//        () ->
+//            assertEquals(
+//                "https://opensource-demo.orangehrmlive.com/web/index.php/auth/login",
+//                driver.getCurrentUrl(),
+//                "URL doesn't match"));
+//  }
 }
